Fix line item dissapearing

+++ src/controllers/api/v1/CartsController.php
@@ -272,10 +272,11 @@ class CartsController extends BaseApiController
                 throw new ValidationException($errors);
             }
 
-            // Since we are replacing all line items (not updating) we could be breaking uniqueness validation on the orders line items.
-            // We need to delete the line items in the database to stop the error.
-            $commerce->getLineItems()->deleteAllLineItemsByOrderId($cart->id);
-
+            if (isset($payload->items)) {
+                // Since we are replacing all line items (not updating) we could be breaking uniqueness validation on the orders line items.
+                // We need to delete the line items in the database to stop the error.
+                $commerce->getLineItems()->deleteAllLineItemsByOrderId($cart->id);
+            }
             // save the cart
             if (!Craft::$app->getElements()->saveElement($cart)) {
                 throw new Exception('Could not save the cart: '.implode(', ', $cart->getErrorSummary(true)));
